refactor: Height 원시값 포장 진행
